api/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
mangum
pytest
python-dotenv
google-generativeai
httpx
